[strategies] Add getters for expand box strategies in envelope box strategies.
