+++ include/boost/geometry/strategies/spherical/envelope_box.hpp
@@ -18,23 +18,8 @@
 #ifndef BOOST_GEOMETRY_STRATEGIES_SPHERICAL_ENVELOPE_BOX_HPP
 #define BOOST_GEOMETRY_STRATEGIES_SPHERICAL_ENVELOPE_BOX_HPP
 
-#include <cstddef>
 
-#include <boost/geometry/core/cs.hpp>
-#include <boost/geometry/core/coordinate_dimension.hpp>
-#include <boost/geometry/core/coordinate_system.hpp>
-#include <boost/geometry/core/tags.hpp>
-
-#include <boost/geometry/views/detail/indexed_point_view.hpp>
-
-#include <boost/geometry/algorithms/convert.hpp>
-#include <boost/geometry/algorithms/detail/convert_point_to_point.hpp>
-#include <boost/geometry/algorithms/detail/normalize.hpp>
-#include <boost/geometry/algorithms/detail/envelope/transform_units.hpp>
-
-#include <boost/geometry/algorithms/dispatch/envelope.hpp>
-
-#include <boost/geometry/geometries/helper_geometry.hpp>
+#include <boost/geometry/strategies/spherical/expand_box.hpp>
 
 #include <boost/geometry/strategies/envelope.hpp>
 
@@ -42,73 +27,18 @@
 namespace boost { namespace geometry
 {
 
-#ifndef DOXYGEN_NO_DETAIL
-namespace detail { namespace envelope
-{
-
-
-template
-<
-    std::size_t Index,
-    std::size_t DimensionCount
->
-struct envelope_indexed_box_on_spheroid
-{
-    template <typename BoxIn, typename BoxOut>
-    static inline void apply(BoxIn const& box_in, BoxOut& mbr)
-    {
-        // transform() does not work with boxes of dimension higher
-        // than 2; to account for such boxes we transform the min/max
-        // points of the boxes using the indexed_point_view
-        detail::indexed_point_view<BoxIn const, Index> box_in_corner(box_in);
-        detail::indexed_point_view<BoxOut, Index> mbr_corner(mbr);
-
-        // first transform the units
-        transform_units(box_in_corner, mbr_corner);
-
-        // now transform the remaining coordinates
-        detail::conversion::point_to_point
-            <
-                detail::indexed_point_view<BoxIn const, Index>,
-                detail::indexed_point_view<BoxOut, Index>,
-                2,
-                DimensionCount
-            >::apply(box_in_corner, mbr_corner);
-    }
-};
-
-
-}} // namespace detail::envelope
-#endif // DOXYGEN_NO_DETAIL
-
-
 namespace strategy { namespace envelope
 {
 
 
 struct spherical_box
+    : geometry::detail::envelope::envelope_box_on_spheroid
 {
-    template <typename BoxIn, typename BoxOut>
-    static inline void apply(BoxIn const& box_in, BoxOut& mbr)
+    typedef typename strategy::expand::spherical_box box_expand_strategy_type;
+
+    static inline box_expand_strategy_type get_box_expand_strategy()
     {
-        // BoxIn can be non-mutable
-        typename helper_geometry<BoxIn>::type box_in_normalized;
-        geometry::convert(box_in, box_in_normalized);
-        
-        if (! is_inverse_spheroidal_coordinates(box_in))
-        {
-            strategy::normalize::spherical_box::apply(box_in, box_in_normalized);
-        }
-
-        geometry::detail::envelope::envelope_indexed_box_on_spheroid
-            <
-                min_corner, dimension<BoxIn>::value
-            >::apply(box_in_normalized, mbr);
-
-        geometry::detail::envelope::envelope_indexed_box_on_spheroid
-            <
-                max_corner, dimension<BoxIn>::value
-            >::apply(box_in_normalized, mbr);
+        return box_expand_strategy_type();
     }
 };
